# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "API to manage a Mechanic Shop"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
...
  /mechanics/{id}:
    get:
      summary: Get mechanic by ID
      tags:
      - Mechanics
      parameters:
      - name: id
        in: path
        required: true
        type: integer
      responses:
        '200':
          description: Mechanic object
          schema:
            type: object
    put:
      summary: Update mechanic
      tags:
      - Mechanics
      parameters:
      - name: id
        in: path
        required: true
        type: integer
      - in: body
        name: body
        required: true
        schema:
          type: object
          properties:
            name:
              type: string
            specialty:
              type: string
      responses:
        '200':
          description: Mechanic updated
      security:
      - Bearer: []
    delete:
      summary: Delete mechanic
      tags:
      - Mechanics
      parameters:
      - name: id
        in: path
        required: true
        type: integer
      responses:
        '200':
          description: Mechanic deleted
      security:
      - Bearer: []
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: 'JWT Authorization header using the Bearer scheme.'
...